P0.9: ревью-цикл исполнен — 18 замечаний трёх оптик внесено: слоты Addr (-1..3), матрица (encode→crypto, loginlink/admin→transport, protocol→«—», +login), реестр (+advisory, +SoA-свап), net→conn, svc→service, интерфейсы сокращены до двух (Handler/Advisory), пакеты loginlink/admin удалены (строки-заготовки), критерий жизни скелета, checkdeps: -test+присваивания, gofmt-гейт, template.md без отсылок, red-proof двух видов

+++ internal/encode/encode.go
@@ -6,6 +6,6 @@ package encode
 // Frame — исходящий кадр клиента в пер-клиентском FIFO (payload уже собран
 // пакетом-писателем).
 type Frame struct {
-	ClientID uint32
+	ClientID uint64 // идентификатор коннекта клиента
 	Payload  []byte
 }
